// src/types/IFormValues.tsx
export default interface IFormValues {
  b: string | number;
  h: string | number;
  d: string | number;
  fck: string | number;
  fyk: string | number;
  E: string | number;
  qsic: string | number;
  qsis: string | number;
  qsif: string | number;
  serviceBendingMoment: string | number;
  beta: string | number
}